+++ lib/routes/create_watcher.dart
@@ -2,8 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_form_builder/flutter_form_builder.dart';
 import 'package:form_builder_validators/form_builder_validators.dart';
 
-import '../widgets/form_field_item.dart';
-import '../widgets/form_widget.dart';
+import '../widgets/form_widget/form_field_item.dart';
+import '../widgets/form_widget/form_widget.dart';
 
 class CreateWatcherRoute extends StatefulWidget {
   const CreateWatcherRoute({Key? key, required this.title}) : super(key: key);
