+++ test/form_widdget_test.dart
@@ -1,11 +1,11 @@
-import 'package:cabinet/widgets/form_field_item.dart';
-import 'package:cabinet/widgets/list_input/select_list_input.dart';
-import 'package:cabinet/widgets/list_input/text_list_input.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_form_builder/flutter_form_builder.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import 'package:cabinet/widgets/form_widget.dart';
+import 'package:cabinet/widgets/form_widget/form_widget.dart';
+import 'package:cabinet/widgets/form_widget/form_field_item.dart';
+import 'package:cabinet/widgets/form_widget/list_input/select_list_input.dart';
+import 'package:cabinet/widgets/form_widget/list_input/text_list_input.dart';
 
 void main() {
   testWidgets("should render empty form widget if no groups are provided",
@@ -13,7 +13,7 @@ void main() {
     await tester.pumpWidget(MaterialApp(
         home: Scaffold(
             body: FormWidget(
-      groups: [],
+      groups: const [],
       formKey: GlobalKey<FormBuilderState>(),
     ))));
